requirements/requirements.txt
fastapi==0.100.0b1
pydantic==2.0b3
uvicorn==0.22.0
pytest=7.3.2
httpx==0.24.1
openai==0.27.8
requests=2.31.0
python-dotenv==1.0.0